+++ BusinessLogic/Service/Application/LocationService.cs
@@ -19,6 +19,11 @@ namespace BusinessLogic.Service.Application
             iLocationRepository = _iLocationRepository;
         }
 
+        public List<Location> Get()
+        {
+            return iLocationRepository.Get();
+        }
+
         public Location Get(int id)
         {
             return iLocationRepository.Get(id);
@@ -42,17 +47,26 @@ namespace BusinessLogic.Service.Application
 
         public bool Update(int id, LocationVM locationVM)
         {
-            return iLocationRepository.Update(id, locationVM);
+            if (string.IsNullOrWhiteSpace(locationVM.Id.ToString()))
+            {
+                return status;
+            }
+            else
+            {
+                return iLocationRepository.Update(id, locationVM);
+            }
         }
 
         public bool Delete(int id)
         {
-            return iLocationRepository.Delete(id);
-        }
-
-        public List<Location> Get()
-        {
-            return iLocationRepository.Get();
+            if (string.IsNullOrWhiteSpace(id.ToString()))
+            {
+                return status;
+            }
+            else
+            {
+                return iLocationRepository.Delete(id);
+            }
         }
     }
 }
